Removes the inflection of the acronym "API"
Host applications that have their own API controllers may not have
chosen to inflect "api" to "API"

In that scenario, simply including the CASino Engine into that application
will cause the entire app to "blow up".

In accordance with the "Principle of least astonishment", CASino should not
mandate how certain words are inflected unless they are entirely unique to
CASino.

* Removes the inflection rule for the API acronym

* Renames the few classes/modules that live within that namespace

* Aliases the "Api" namespace to "API" for applications that *have* chosen
  to inflect "api" to "API"