Reproduces the original CASinoCore schema
This allows `casino:install:migrations` to work for both CASino 1.x as well
as completely clean installs.

Under a clean install, this new migration will create all of the CASino
database tables with the older CASinoCore names. These will then be renamed
by the upgrading migration.

Under a CASino 1.x install, this new migration will do nothing because the
tables already exist. The upgrade migration will then handle renaming all
of the tables.

When rolling back, the upgrade migration will go ahead and drop all of the
CASino tables and indices. This is because there is no programmatic way to
determine whether the migration ran with the context of a prior install or
a clean install.